move LuaImplPackage into the lua package
