+++ lib/spack/spack/pkgkit.py
@@ -26,7 +26,6 @@ from spack.build_systems.cuda import CudaPackage
 from spack.build_systems.gnu import GNUMirrorPackage
 from spack.build_systems.intel import IntelPackage
 from spack.build_systems.lua import LuaPackage
-from spack.build_systems.lua_impl import LuaImplPackage
 from spack.build_systems.makefile import MakefilePackage
 from spack.build_systems.maven import MavenPackage
 from spack.build_systems.meson import MesonPackage
